Fix sonar cloud issue
